Refactor/reorganize streaming code
